+++ PAC/linux/l_tcp_cmctr.h
@@ -68,7 +68,7 @@ class tcp_communicator_linux : public tcp_communicator
             /// @param skt - сокет.
             int do_echo( int idx );
 
-            u_long glob_last_transfer_time;  ///< Время последней успешной передачи данных.
+            uint32_t glob_last_transfer_time;  ///< Время последней успешной передачи данных.
 
             timeval tv;                      ///< Задержка ожидания функции опроса состояний сокетов, 0 по умолчанию.
             fd_set rfds;                     ///< Набор дескрипторов сокетных файлов для чтения.
