+++ PAC/win/w_tcp_client.h
@@ -8,11 +8,11 @@ class win_tcp_client: public tcp_client
 	private:
 		timeval tv;
 		fd_set rfds;
-		unsigned long async_startconnnect;
+		uint32_t async_startconnnect;
 		timeval async_tv;
 		sockaddr_in async_sock_address;
 		unsigned int async_bytes_to_send;
-		u_long asyncReceiveTime;
+		uint32_t asyncReceiveTime;
 		bool isNewData;
 	public:
 		int InitLib();
@@ -27,7 +27,7 @@ class win_tcp_client: public tcp_client
 
 		int AsyncReceive() override;
 		virtual int get_async_result() override;
-		win_tcp_client( const char* client_ip, unsigned int client_port, unsigned int client_id, unsigned char alarm_subclass, unsigned int exchange_buf_size = 256, unsigned long send_receive_timeout = 100 );
+		win_tcp_client( const char* client_ip, unsigned int client_port, unsigned int client_id, unsigned char alarm_subclass, unsigned int exchange_buf_size = 256, uint32_t send_receive_timeout = 100 );
 
 		~win_tcp_client();
